+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.7] - 2021-06-09 🍉
+## [1.0.7] - 2021-06-11 🍉
 - Adds a `TestClient` class that simplifies testing of applications
 - Fixes bug [#156](https://github.com/Neoteroi/BlackSheep/issues/156),
   preventing route parameters to work when the user doesn't follow Python
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds support for automatic generation of OpenAPI Documentation for `Generic`
   types
 - Improves the generation of OpenAPI Documentation for `pydantic` types and to
-  support more object types
+  support more object types (fixes [#167](https://github.com/Neoteroi/BlackSheep/issues/167))
+- Ensures that request body is parsed as JSON only if the content type contains
+  the "json" substring
 
 ## [1.0.6] - 2021-05-30 :birthday:
 - Fixes bug [#153](https://github.com/Neoteroi/BlackSheep/issues/153),